scraper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "visquery-scraper"
version = "0.1.0"
description = "Scrapy-based scraper subsystem for Visquery architectural precedent search"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "scrapy>=2.11",
    "playwright>=1.44",
    "scrapy-playwright>=0.0.34",
    "Pillow>=10.3",
    "imagehash>=4.3",
    "psycopg2-binary>=2.9",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "lxml>=5.2",
    "w3lib>=2.1",
    "asyncpg>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "scrapy-fake-useragent>=1.4",
]

[tool.hatch.build.targets.wheel]
packages = ["visquery_scraper"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
